Add files via upload
 My team has made a webpage and an android application which aims to manage
evidences using blockchain technology. According to a survey 3 out of 4 cases go unreported
in India & with the help of our app common people will be the eyes of the police department
and will be able to report crimes easily through their mobile phones or with the help of a
computer , their reports will be verified as legitimate or fake with the help of an AI module and
after the verification of the complaint filed the local police authority will be notified so that
they can take action against the crime reported and As a token of appreciation the person who
had reported the crime will get a gift voucher from the government which he/she can redeem
anytime or anywhere they want and the whole process will be secured with the help of
blockchain technology.